refactor: Drawer 스타일을 공통 스타일로 분리
